pyproject.toml
[project]
name = "starreach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
]
